Added new features to working with the queue
Now you can mark whether a student's work has been accepted or not, and also send it to the back of the queue. Added a display of the number of students in the queue and the time until the end of retake session. Added the ability to extend the retake time (a window pops up) and close it. And much more...